Move a bunch of files around to improve general structure
