--- src/setupTests.tsx
@@ -1,4 +1,0 @@
-import '@testing-library/jest-dom/extend-expect';
-import 'mutationobserver-shim';
-
-global.scroll = function() {};
