+++ src/ui/settings/language.tsx
@@ -4,7 +4,7 @@
 import { jsx } from 'theme-ui';
 import { useTranslation } from 'react-i18next';
 
-import languages from '../../languages';
+import languages from '../../i18n/languages';
 import { SettingsSection } from './elements';
 
 const LanguageSettings = () => {
