+++ src/ui/studio/recording/recorder.tsx
@@ -1,5 +1,6 @@
-import { Settings } from './settings';
-import { dimensionsOf } from './util';
+import { Settings } from '../../../settings';
+import { dimensionsOf } from '../../../util';
+
 
 export type OnStopCallback = (args: {
   url: string,
